pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = "A music platform for artists to publish their albums."
authors = ["AmmarDab3an"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-model-utils = "^4.3.1"
django-debug-toolbar = "^3.8.1"
django-imagekit = "^4.1.0"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
